Introduce memory-revoking-target concept in PoS revoking logic
Currently, when revoking is triggered, everything is spilled to
disk. This is a bit aggresive since we can keep some data in
memory itself and only dump a portion to disk so that enough
memory is released for the query to continue. This PR introduces
a memory-revoking-target config which can be used to specifiy
how much data should be spilled on revoke.