Decouple feature project and struct encoding option (#826)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/826

# Problem
`FlatMapColumnReader` supports reading a map column as a struct encoding with a given list of keys. However, it was using key filtering expression through `ColumnSelector`, as a way to pass the key list. This was working until we expanded the expression to support reject lists.

Now the expression cannot represent a list of keys to accept. We need an explicit way to pass a list of keys to construct a struct encoding, separately from the key filtering.

This is required to have a consistent output schema even with key filtering.

# Solution
I updated `RowReaderOption` to not only accept a list of nodes, but also a list of keys for each configured node. Now we can configure it to produce N fields struct but reading only a subset of them.

Reviewed By: zzhao0

Differential Revision: D33352676

fbshipit-source-id: 4cab57687a1ef6408b2eb293de987d46ffeb8d21