Simplify FusedLoad and CachedBufferedInput (#840)
Summary:
The IO pattern for Velox file readers is to register the pieces of file to be
accessed, then to access some or all of these pieces. Pieces that are in close
proximity get loaded with a single IO. The definition of close depends on the
media.

This removes the load_ member from AsyncDataCacheEntry. The cache entry no
longer references the FusedLoad that will be invoked to fill it. Instead, the
CacheInputStream  that will load a piece of file coalesced with other pieces
references the FusedLoad.

The API contract of CachePin is simplified: If findOrCreate returns an exclusive
pin, the caller is responsible for filling the entry and calling
setExclusiveToShared when done. This makes the data visible to other threads.
There is no more a concept of dataValid or ensureLoaded for shared mode pins
that have uninitialized data.

We add tests for checking a consistent after state for errors in loading data
into CachePins. AsyncDataCacheTest adds error injection in FusedLoads and
single pin loading. The code is also more representative of the intended usage
pattern.

CachedBufferedInput is adapted to use the changed AsyncDataCache API.
CacheInputStream looks for its FusedLoad instead of the FusedLoad being
recorded in a cache entry. Cache entries for FusedLoads are made right before
loading the FusedLoad. CoalesceIO.h is used systematically to detect and
coalesce coalesced loads.

ScanTracker exports TrackingData so that columns can be more accurately sorted
in different load frequency buckets.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/840

Reviewed By: mbasmanova

Differential Revision: D33510308

Pulled By: oerling

fbshipit-source-id: be686bbba41cbacf1a5f90e04063b12b97007746