Fix Expr bug on peeling encoding for single-arg deterministic func (#866)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/866

Motivation:
A crash was hit when running expression `not(contains(...))` and the root cause is nullptr reference in casting the input to flat vector. This indicates that a bug on argument encoding peeling of `Expr`.

Fix:
The reproduce scenario is turned into a test case of ExprTest where the input is dictionary vector but with `DictionaryVector::isConstant(rows)` equals to true. We handled this by creating a ConstantVector of out it so that the further execution can go into `applySingleConstArgVectorFunction`.

Reviewed By: mbasmanova

Differential Revision: D33523786

fbshipit-source-id: 693cc04d4f0617e7639206d5dd6a41634451d150