Improve error message in evaluate() to show result vector encoding
Summary: `evaluate()` attempts to dynamically cast the result vector pointer to a specified target vector type. When the dynamic pointer cast fails, it emits an error message with the actual vector element type included as a hint for developers. However, there are cases where the dynamic pointer cast fails because of a mismatch of the vector type itself instead of the vector element type. An example is when the output of a UDF call is expected to be an array vector of one single row wrapped in dictionary encoding, the result actually becomes a constant vector of array type. In such a case, the dynamic pointer cast in `evaluate()` fails even though the vector element type matches with the expectation. Therefore, this diff adds the vector type to the error message to give developers more complete information.

Reviewed By: mbasmanova

Differential Revision: D33557516

fbshipit-source-id: ce9ef3929abee55756fcc49ff9a376d5b618aec7