Disabling flaky test
It seems the query order is not respected all the time due to the nature of multi threading with Dispatcher logic.
So the first query submitted might end up getting queued and the second one runs if the thread handling second one gets prioritized.
This leads to the failure, needs better investigation to see what should be the right fix here. Disabling till then.