Fix flaky cache tests (#870)
Summary:
Join with executor for async operations before teardown of test. Fixes
possible access to freed memory.

Use MmapAllocator for test in order to have a hard cap on allocation
volume.

Fix accounting of cached pages for the case of removing a cache
entry after failure to fill it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/870

Reviewed By: mbasmanova

Differential Revision: D33539438

Pulled By: oerling

fbshipit-source-id: a7678c79a468189dad2e58644e27711ecea867ec