Fix count(*) query over Parquet files (#893)
Summary:
As reported in https://github.com/facebookexternal/presto_cpp/issues/616, a `select count(*) from t` query over Parquet files fails with an error `failed: !filter_.empty() need a valid filter`.

Fix by removing the check in the `ColumnSelector::getProjection()` method that verifies that `ColumnFilter` is not empty. It is possible that the `ColumnFilter` is empty as in the `select count(*)` case.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/893

Reviewed By: zzhao0

Differential Revision: D33663775

Pulled By: mbasmanova

fbshipit-source-id: f0b426aa135859c583ca10c5f9112485c9ba0083