Ensure evaluation always returns a valid vector. (#913)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/913

Previously evaluating an expression with an empty vector / no
input rows doesn't populate result, leading to an easy source of SEGV
unless every client of Velox explicitly checks for, and works around
this. Here we simplify the client logic and make it safer by returning
an empty vector of the correct type rather than a null result.

A great live example of a SEGV in practice:
https://github.com/facebookresearch/torcharrow/issues/35

Reviewed By: pedroerp

Differential Revision: D33716916

fbshipit-source-id: a3a767d4baacc22c8a689de374d0821d0e3a5e50