Add an SSD layer for AsyncDataCache (#488)
Summary:
If AsyncDataCache is given an optional SsdCache, moves qualifying
entries into the SsdCache as a by-product of scanning entries for
eviction.

SsdCache has the same key and quantization as the RAM cache. Writing
to SSD is done on a best effort basis so as not to block reading from
storage while waiting to save on SSD.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/488

Reviewed By: mbasmanova

Differential Revision: D33686894

Pulled By: oerling

fbshipit-source-id: 250aa923ce75a66eb08ceebeb29759a511406312