Fix cardinality of Boolean in VectorHasher (#921)
Summary:
Fix a bug with array/normalized key aggregation with boolean keys.

When the data is all false for the creation of the hash table, the
column has a cardinality of 1 by max-min. When a true value comes
along, this is mapped to a constant without looking at the min-max and
indices go out of range. So a boolean key will have a cardinality of 3
for either range or value ids. A boolean key is always expected to
have a range mapping. cardinality() was correct but enableValueRange()
was not consistent with it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/921

Reviewed By: mbasmanova

Differential Revision: D33767021

Pulled By: oerling

fbshipit-source-id: 27975d21700bf10d9f91f9576556bab5e6398785