Fix Parquet reader filter index (#873)
Summary:
The filter column Id passed to DuckDB Parquet reader should be the
relative index to the projection list. If a filter column is not
included in the projection list, it is still included in the columnIds
array and the output from the reader will include this column. It will
be discarded in HiveConnector later when outputColumns is being built.

Resolves https://github.com/facebookincubator/velox/issues/865

Pull Request resolved: https://github.com/facebookincubator/velox/pull/873

Reviewed By: pedroerp

Differential Revision: D33768381

Pulled By: mbasmanova

fbshipit-source-id: 49840a2edaf87610e26caa1c5269ab5a2d107929