Implement grouped execution (#914)
Summary:
The first cut of support for grouped execution, when we receive splits in split groups and want to run each group separately to keep hashmaps smaller as in that cases the hash keys are guaranteed not to intersect between any two groups.

The change is somewhat larger than I would prefer due to me having little idea about how things worked prior to this diff and even now. :)

I will comment in one of the later versions of the diff why I did this or that.

There are no tests at the moment, except for single e2e test I've been running locally.
We'll hav to think what tests we want to subject this new functionality to.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/914

Reviewed By: mbasmanova

Differential Revision: D33725845

Pulled By: spershin

fbshipit-source-id: f118634009b4e3a445104db4b0f14e1051495807