Add mayHaveNullsRecursive function to Vectors (#938)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/938

# High Level Goal:
We've identified some use cases where they want UDFs to return NULL not only if any of the values of the arguments to a function are NULL, but also if any of those arguments are complex types containing NULLs.

In these cases these UDFs they want a variant of the call function which receive views that don't expose nullity (since there are guaranteed to be no NULLs).  E.g. for Array<int64_t> instead of getting an ArrayView that offers an interface similar to std::vector<std::optional<int64_t>, they want an ArrayView that offers an interface like std::vector<int64_t>.

I plan to augment the SimpleFunction interface to offer a callNullFree function, which is only called if the arguments are not NULL and do not contain NULLs.  This function exposes Views with the desired interface as described above.  If only callNullFree is implemented, SimpleFunctions will return NULL in the if callNullFree cannot be invoked, otherwise it acts like a fast path if we can quickly determine the inputs do not contain NULLs and otherwise we defer to call or callNullable.

# This Diff
I add a function mayHaveNullsRecursive to Vectors which returns false iff we can be sure that the vector and all of its children have nullptr for nulls_ (or isNull_ == true for constants).  This gives us a quick check we'll be able to apply in the SimpleFunctionAdapter to determine if we can call the callNullFree function, or if we need to do a slower thorough search for nulls if only callNullFree is implemented.

Reviewed By: laithsakka

Differential Revision: D33829875

fbshipit-source-id: fb513e9256fbc123ead8c058abacf7dcf1d18ba2