Add AsyncSource template for speculative execution (#903)
Summary:
AsyncSource encapsulates a background computation. These can be
scheduled on a background executor. The difference between this and a
future is that when the user requires the result, AsyncSource will
perform the async peration on the caller's thread and will turn the
background operation into a no-op.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/903

Reviewed By: pedroerp

Differential Revision: D33687621

Pulled By: oerling

fbshipit-source-id: 46f3e7613bd67bc67e7b43336fc64e0d74c9c8ba