Release dictionary memory on Stripe flush
1. When stripe is flushed, OrcOutputBuffer holds the allocated memory.
This memory is unused until the next stripe flush. This memory can be
comparable in size to the dictionary memory, hence doubling the memory
required by the writer for the subsequent stripe flushes. Now
OrcOutputBuffer is reallocated for each stripe. Further as soon as the
dictionary is not required they are set to null.
2. When dictionar is abandoned, the hash buckets are not freed up.
This can be few megabytes per column and for multiple columns this
adds up. Now closeDictionary frees up the dictionary completely.

This change is a trade off, it reduces the memory required by writer
by 20% in most cases, while having marginal CPU impact.