Add GenericView, and VectorReader<Generic> (#957)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/957

This diff adds the Generic static type, and GenericView
which represents that type when reading from VectorReader<Generic>
which was also added in this diff.

- GenericView support two operations (compare and hash), in the future I
am planning to add casting to other types(unchecked and dynamic safe)
 and ability to access the type.
- While this is not plugged into the simple function interface yet. It
can be used in the vector functions to simplify implementations
of functions such as array_union and array_position.

- The diff tests VectorReader<Generic> with int, and array<int> .
(The generic is int, or array<int>)
- The diff also tests a VectorReader<ArrayView<Generic>>
with an array<int>. (generics are ints).

- Next, I will add more testing to test different possible types,
although the logic is the same and not dependent on the type.

- Also it might be useful to add compare and hash functionality to
ArrayView and MapView.

- Finally, I will have to plug this in the simple function interface and its
registration system.

**Note on possible optimization:**
I was thinking that if we allow multiple functions to be registered with
same name but different type signatures and carefully do the resolution of function
by prioritizing non-generic signatures first we can register generic functions with an enabled fast path for non-generic

for example, a function compare is simple
```
template<A,B>
Call (){out = a==b};
```

can be registered for all primitives and for generic.
```
register<cmp, bool , int, int>();
register<cmp, bool , Generic, Generic>();
```
during resolutions two functions will be available for an int
input, but the first should be picked for that . And the one
with Generic would be the last option to pick as a the
general path of nested complex types. or complex types in
general.

Reviewed By: funrollloops

Differential Revision: D33809195

fbshipit-source-id: 44e71fd210f6df6a7d2fe11fa1a7e0dde7d9668b