Add Non Nullable variants of Views used in Simple Functions (#926)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/926

# High Level Goal:
We've identified some use cases where they want UDFs to return NULL not only if any of the values of the arguments to a function are NULL, but also if any of those arguments are complex types containing NULLs.

In these cases these UDFs they want a variant of the call function which receive views that don't expose nullity (since there are guaranteed to be no NULLs).  E.g. for Array<int64_t> instead of getting an ArrayView that offers an interface similar to std::vector<std::optional<int64_t>, they want an ArrayView that offers an interface like std::vector<int64_t>.

I plan to augment the SimpleFunction interface to offer a callNoNulls function, which is only called if the arguments are not NULL and do not contain NULLs.  This function exposes Views with the desired interface as described above.  If only callNoNulls is implemented, SimpleFunctions will return NULL in the if callNoNulls cannot be invoked, otherwise it acts like a fast path if we can quickly determine the inputs do not contain NULLs and otherwise we defer to call or callNullable.

# This Diff

This diff adds a bool to the template arguments of the Views used for reading complex types in Simple Functions.  This new template argument indicates whether or not the values exposed by this View can be NULL.

If set to false, the View will return the underlying object directly, rather than wrapping it in a VectorOptionalValueAccessor object when functions like at or the [] operator are called.  Likewise the hasValue function in the Iterators and the mayHaveNulls functions are modified to invariably return true or false respectively.

I also modified the Reader interface to add the function readNullFree which returns the new null free Views for complex types.

Lastly, I added an in_type_null_free to resolvers for complex types which refers to the null free version of the container's View.

Reviewed By: laithsakka

Differential Revision: D33778701

fbshipit-source-id: ceb37817b1eb3f478ced4d85c2c4c3300178243e