Fix deadlock in Task::removeDriver (#951)
Summary:
Fix deadlocks in Driver/Task termination

A task can be terminated anytime by a Driver inside it or any other thread.

Drivers are created as part of a Task. Before they can be destroyed,
they have to be unhooked from the Task. Otherwise the Task has a
shared_ptr to the Driver and the Driver to the Task and both are live
indefinitely.

Destruction of a Driver should not reference the Task. This is a
liability for deadlock if the last reference is dropped inside
Task::mutex_.

The Task termination is split into off-thread Drivers and on-thread
Drivers. The off-thread Drivers are unhooked from the Task inside
Task::mutex_. The on-thread ones will unhook themselves when they
notice the Task's terminateRequested_ flag, at latest when they go off
thread.

Updates and fixes some comments.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/951

Reviewed By: spershin

Differential Revision: D33881765

Pulled By: oerling

fbshipit-source-id: 6900dd83b90ef75840ff82593bca2aa7d1602ec9