support accumulator using non Velox managed memory (#930)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/930

Motivation:
We have a set of Aggregates that use accumulator that is hard to migrate to Velox managed memory. In order to run them in Velox, we decided to add support for accumulator using non-Velox managed memory.

Design:
1.Velox provides an interface to initialize and free accumulator using external memory: initializeNewGroups and destroy(...).

2.Velox calls initialize method to “create”/”copy”/”move” accumulator and destroy method to free accumulator's memory.

Velox Aggregate only does “create” and “free” of the accumulator, it never “move” or “copy”. The "create" and "free" are done through `velox::exec::RowContainer`.

For “create”, Velox already calls initializeNewGroups(...) every time a new accumulator is created.

For “destroy”, Velox already calls destroy(...) when a row is erased or reused. But it does not call destroy(...) when we clear everything, so this is what we need to add. We also add `accumulatorUsesExternalMemory()` to velox Aggregate so we will only call destroy(...) if necessary.

Concern:
One concern about using external memory is that Velox won't be able to tell how much of the memory is used by the query and won't be able to use spilling. The application would need to monitor the total memory and take actions as necessary.

Reviewed By: mbasmanova

Differential Revision: D33716974

fbshipit-source-id: 8824081e763c42c47db033c378086ba2182b4120