Fix read bytes accounting (#943)
Summary:
When using CachedBufferedInput, do not have an IoStatistics in the InputStream
passed to the CachedBufferedInput. The CachedBufferedInput records the IO
directly. This changes the pattern where InputStream records IO and
CachedBufferedInput/CacheInputStream then reverts the increment to avoid double
counting.

Compensating for double counting can lead to negative counts with multithreaded
reading, which in turn kills the query with Presto when the coordinator errors
out from having a negative size.

Test the invariant relations between counters:

- rawBytesRead is what the reading threads accesses. This is less or equal to
  what is hit from RAM + what is read from storage + what is read from SSD.

- read is what is read from storage.

- ssdRead is what is read from SsdCache.

- prefetch is what is speculatively read from storage or SSD. This is less than
  the sum of SSD and storage reads since not all read is speculative.

- ramHit is what is accessed and is already in RAM. This is less or equal to
  what is accessed. Data that is fetched from SSD or storage with prefetch and
  then hit does not count as ramHit on the first use.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/943

Reviewed By: mbasmanova

Differential Revision: D33854323

Pulled By: oerling

fbshipit-source-id: 26b7a0c17877704b3519de294447265722b4033a