Fix flaky DriverTest.pause (#982)
Summary:
The test checks that numRunningDrivers is 0 after cancellation is
synced using finishFuture(). The Drivers are no longer on thread when
the future is realized. They are still in the Task and are in the
process of being closed, which then removes them from the Task.

We make numRunningDrivers_'' 0 on the first call to terminate and do
not further decrement this on Driver closing if the task is not in a
running state.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/982

Reviewed By: spershin

Differential Revision: D33993781

Pulled By: oerling

fbshipit-source-id: 434baf3f6896614fdafe6df4eea5eedbd7afce6a