Add preGroupedKeys field to Aggregation plan node (#988)
Summary:
Remove StreamingAggregationNode and add preGroupedKeys field to AggregationNode
instead. Update LocalPlanner to use StreamingAggregation operator if input if
pre-grouped / clustered on all grouping keys. A follow-up PR will introduce an
optimization in HashAggregation operator for the case when input is clustered
on some, but not all grouping keys. In this case, we can reduce memory
footprint of the aggregation by flushing accumulators and the hash table every
time one of the pre-grouped keys value changes.

We can also optimize hashing of the high-cardinality pre-grouped keys by mapping
their values to a small integer. We would iterate over the input vector and
assign repeated sequential numbers to each row. We would start with 0 and
increment the number every time one of the grouping keys change. This way, all
pre-grouped keys will map to a small integer domain and allow for array-based
aggregation if the other keys also have low cardinality.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/988

Reviewed By: laithsakka

Differential Revision: D34005689

Pulled By: mbasmanova

fbshipit-source-id: 0e06780cab12079a763024b76292bb44c877384d