Fix flaky IO count (#991)
Summary:
The assertion that total bytes accessed <= hits from RAM + read from
SSD + read from storage is not always true at the level of the
thread. It is possibl that both the read and the hit get counted on
the same thread if another thread accesses an entry prefetched on
another thread. Then the thread that hits the entry after the prefetch
but before the prefetcher initiating thread does not count the
hit. This is a marginal edge case that will not significantly alter
stats.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/991

Reviewed By: kagamiori

Differential Revision: D34020670

Pulled By: oerling

fbshipit-source-id: a530c8c7e920e4ac1202aee51322d2b402198cdf