Add flags for enabling AsyncdataCache/SsdCache/MmapAllocator (#630)
Summary:
Introduces flags for enabling the Velox IO subsystem based on
CachedBufferedInput.

--use_mmap_allocator and --use_async_cache should be set
together. Running AsyncDataCache with regular MappedMemory results in
OOM from fragmentation.  When these flags are on,--system_memory_gb
must be specified to set the maximum resident size for large objects,
i.e. cache buffers, hash tables etc. Memory for query plans and
intermediate results is not included in this figure.

-- SSD cache is optional and enabled by a non-0 value of
async_cache_ssd_gb. This is the maximum size of data cache in the
host's local filesystem. If this is set --async_cache_ssd_path must
also be set to point to a writable directory where the cache files
will be written.

Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/630

Reviewed By: spershin

Differential Revision: D33822592

Pulled By: oerling

fbshipit-source-id: dd1858f4433e6349bcebf355e4d89250855af951