Use ContiguousAllocation for hashTable (#952)
Summary:
Uses MappedMemory::ContiguousAllocation for the potentially large
contiguous arrays for HashTable. This replaces bare malloc which was
not accounted for in any tracking. Now the HashTable is accounted for
in the MappedMemory of its RowContainer. This removes fragmentation
and prevents increase of RSS from making large arrays.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/952

Reviewed By: pedroerp

Differential Revision: D33881770

Pulled By: oerling

fbshipit-source-id: 2c5b6ee439941c5a5ca428d505742955679c3289