Test peeling with LazyVectors that produce dictionaries (#995)
Summary:
A LazyVector produces a dictionary for a string dictionary column and
for a sparsely accessed column. In the sparse column case, the values
are contiguous but the dictionary maps them to non-contiguous places
as expected by the call site.

When a LazyVector is an argument to a function, it gets loaded,
becomes a dictionary and is then peeled. After this, the function can
return a constant, which is then re-wrapped in the dictionary. This
gets peeled again by the next outer function, which should run on a
single value since the dictionary is over a constant. We fix the
detection of all constant arguments in applyFunctionWithPeeling. An
argument can start as a wrapper and become a constant, which
possibility was not considered, leading to failing the check for
single active row in Expr::applySingleConstArgVectorFunction.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/995

Reviewed By: mbasmanova

Differential Revision: D34048224

Pulled By: oerling

fbshipit-source-id: 8778b2ac0835d64d28525ccef7954d347ad5a3e9