Option to disable string dictionary encoding in OrcWriter
There are tables with 4000+ string columns. Each String dictionary
column takes about 100 KB when initialized (40KB for row group index),
32KB for SegementedSliceBuilder and few others. So they all together
combine to make the default writer start with 700MB+ and cause it to
OOM. This change provides an option to disable the String dictionary
encoding alltogether in the ORC writer.