Optimize aggregation over partially clustered input (#984)
Summary:
Optimize aggregation with multiple grouping keys for the case when input is
pre-grouped / clustered on a subset of these keys. In this case, there is no
need to accumulate all the input in memory before producing results. We can
produce results and flush memory every time a value of on the of pre-grouped
keys changes. This allows to reduce memory footprint of the aggregation.

A couple of straightforward optimizations can be added in follow up PRs.

First, we can add logic to avoid emitting very small batches of output by
waiting to accumulate at least N groups before producing output.

Second, we can optimize hashing of the high-cardinality pre-grouped keys by
mapping their values to small integers. We would iterate over the input vector
and assign repeated sequential numbers to each row. We would start with 0 and
increment the number every time one of the grouping keys change. This way, all
pre-grouped keys will map to a small integer domain and allow for array-based
aggregation if the other keys also have low cardinality.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/984

Reviewed By: oerling

Differential Revision: D34048370

Pulled By: mbasmanova

fbshipit-source-id: 5f163c2e3442a376ac04e032220373431f184326