Add ability to do streaming for partial aggregation
We can always enable streaming aggregation for partial aggregations without affecting correctness.
But if the data isn't clustered (for example: ordered) by the group-by keys, it may cause regressions on latency
and resource usage. This session property is just a solution to force enabling streaming aggregation
when we know the execution would benefit from partial streaming aggregation.
We can work on determining it based on the input table properties later.