Add unit test for grouped/bucketed execution. (#1011)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1011

Adding the unist test to test out-of-sync addition of splits and split groups, also for partial split groups and for adding splits before the Task start.
Added a new Task state (kNone) to reflect that we haven't started yet.
Updated exec task utils classes to be able to work with split groups.

Reviewed By: mbasmanova

Differential Revision: D34088826

fbshipit-source-id: 471a83e96528a003d80d4ed0caa0b930761e4c6d