Fix long dictionary raw bytes estimate
In Long dictionary encoding , unique values are encoded in dictionary
and indexes are stored in data stream. This encoding is beneficial when
the values are large numbers. Long dictionary encoding considered
all values as 8 bytes for raw bytes calculation and this resulted in
small values considered to dictionary stream.

Considered a stream that has only unique values from 1 to 10. Long
Dictionary encoding is not beneficial for these values. If the stream
has 10 unique numbers, but all of them are large numbers then long
dictionary encoding will be beneficial. This PR changes how raw bytes
estimate is calculated for the long dictionary.