First part of performance regression framework (#992)
Summary:
This is the initial part of a performance regression framework for Velox. It adds a few initial tests
based on multiplications on flatVectors, and creates a circleCI job to run it.

This is just the initial steps. The following PRs will add automation to run the tests first again main,
then against the current commit, and compare them.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/992

Reviewed By: laithsakka

Differential Revision: D34182183

Pulled By: pedroerp

fbshipit-source-id: 972424ae9d9a57fb911b46d6b93fdd0b2fd52c9b