Simple functions type analysis. (#1041)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1041

Add an analysis to the simple function types to extract information that is needed
in the next diff to assign priorities to signatures.

The recursion that used to convert the types to their string representation is updated
to a general analysis that collects other information as well.

Unit tests added in the diff.

Reviewed By: kevinwilfong

Differential Revision: D34226886

fbshipit-source-id: c69d46415e9793e2c1c445e9496d5e1dd6d25af3