Refactor null handling in ORC Dictionary writers
Before this change, nulls are tracked in dictionary indexes as negative
indexes. When they are converted to direct encoding or during stripe
flush they are written to PresentOutputStreams. In the final streams
nulls are represented using bit fields but in memory they are represented
as byte, short or int depending on the dictionary size.

Representing them as index had 8,16 or 32(byte, short or int) times overhead.
This caused the writers to OOM as dictionary buffered bytes count only
1/8th of a byte for nulls. The present stream has the same representation
whether it is direct or dictionary encoded data. This change now tracks the
Present stream in dictionary directly and when converting to direct, the
present stream responsibility is transferred to the direct writer.