Add containsNulls functions to VectorReaders (#989)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/989

# High Level Goal:
We've identified some use cases where they want UDFs to return NULL not only if any of
the values of the arguments to a function are NULL, but also if any of those arguments are
complex types containing NULLs.

In these cases these UDFs they want a variant of the call function which receive views that
don't expose nullity (since there are guaranteed to be no NULLs).  E.g. for Array<int64_t>
instead of getting an ArrayView that offers an interface similar to
std::vector<std::optional<int64_t>, they want an ArrayView that offers an interface like
std::vector<int64_t>.

I plan to augment the SimpleFunction interface to offer a callNullFree function, which is only
called if the arguments are not NULL and do not contain NULLs.  This function exposes
Views with the desired interface as described above.  If only callNullFree is implemented,
SimpleFunctions will return NULL in the if callNullFree cannot be invoked, otherwise it acts
like a fast path if we can quickly determine the inputs do not contain NULLs and otherwise
we defer to call or callNullable.

# This Diff
I add a function containsNulls to the VectorReaders, which allows us to inspect the value at
a specific index to see if it is NULL or any of its children (e.g. the elements in an array)
contain NULL.

I add two variants of the function, one that takes a single index, and one that takes a range
of indices, to make evaluating the check for container types easier.

Note in VectorReaders for containers mayHaveNullsRecursive is called on the child vectors
on each call to containsNulls.  This is only temporary, and I plan to compute the value only
once in the next diff because it was difficult to do that in this diff in a way that allowed me
to split the diffs.  (I only want to initialize it when we're going to call containsNulls, and that
context is only available in the next diff).

Reviewed By: laithsakka

Differential Revision: D33921847

fbshipit-source-id: 5db8bc0daba1cc83a9d8d5aeed3336c488755914