Fix flaky tests that use Task operatorStats (#1068)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1068

I've noticed that some tests in LocalPartitionTest and TableScanTest are flaky in Circle CI.

I investigated and it looks like a race condition between when a task is returned by
assertQuery and when the operatorStats are filled in.

The Cursor class is used in tests to execute a query asynchronously.  It uses a queue to
manage running queries.

It uses the ConsumerSupplier interface to enqueue a signal that the query is done
https://github.com/facebookincubator/velox/blob/b242f82aa44f163fa49f05467adbc027ee533014/velox/exec/tests/utils/Cursor.cpp#L132

When the signal is dequeued, assertQuery returns the Task.

The Driver invokes the ConsumerSupplier when an operator can receive no more input
https://github.com/facebookincubator/velox/blob/main/velox/exec/Driver.cpp#L421

However, the OperatorStats for the task aren't populated until the Driver is closed, once all
operators are finished
https://github.com/facebookincubator/velox/blob/main/velox/exec/Driver.cpp#L437

This leads to that race condition.

The fix I've made here is to check to make sure the Task isFinished (or wait for it to be
finished if it's not) before accessing taskStats() and the OperatorStats it contains.  The Task
is only marked finished once the Driver finishes executing.

Reviewed By: mbasmanova

Differential Revision: D34320000

fbshipit-source-id: 8501ee50549069a780704246c5699b3eddf5046f