Add SelectiveTimestampColumnReader (#1037)
Summary:
Adds a timestamp reader and test in E2EFilterTest. There are no
Filters for the timestamp type.

Adds a ColumnVisitor for non-dictionary uses of RLE encoding, which first occurs with the timestamp encoding.

Enhances processFixedWidthRun in DecoderUtil.h to to apply to parts of
the run to decode instead of processing the whole range given at the
top level. This is needed for use in DirectRLEColumnVisitor where this
is called multiple times for repeatied/literal runs.

Adds a test for processFixedRun processing  a column filter in multiple
installments. This covers the changed lines in processFixedWidthRun.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1037

Reviewed By: zzhao0

Differential Revision: D34218687

Pulled By: oerling

fbshipit-source-id: 5b035e897125d9e4b511739ef0c415f258614bd8