Add callNullFree to Simple Function interface (#990)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/990

# High Level Goal:
We've identified some use cases where they want UDFs to return NULL not only if any of the values of the arguments to a function are NULL, but also if any of those arguments are complex types containing NULLs.

In these cases these UDFs they want a variant of the call function which receive views that don't expose nullity (since there are guaranteed to be no NULLs).  E.g. for Array<int64_t> instead of getting an ArrayView that offers an interface similar to std::vector<std::optional<int64_t>, they want an ArrayView that offers an interface like std::vector<int64_t>.

I plan to augment the SimpleFunction interface to offer a callNullFree function, which is only called if the arguments are not NULL and do not contain NULLs.  This function exposes Views with the desired interface as described above.  If only callNullFree is implemented, SimpleFunctions will return NULL in the if callNullFree cannot be invoked, otherwise it acts like a fast path if we can quickly determine the inputs do not contain NULLs and otherwise we defer to call or callNullable.

# This Diff
Here I pull everything together adding the callNullFree function to the Simple Function interface which behaves as described in the High Level Goal.

I've tried to balance writing clean code with writing performant code, tradeoffs were generally made in favor of the latter.

Reviewed By: laithsakka

Differential Revision: D33931112

fbshipit-source-id: bd117a86cbc671f9f500f8761d245f5699b3d11d