ExpressionFuzzer expects TRY to catch exceptions in query compilation (#1078)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1078

Currently when the ExpressionFuzzer is run with the --retry-with-try option, it fails if after
wrapping the expression in a TRY query compilation or execution fails.

It's not possible for TRY to catch exceptions in query compilation, only execution.

Here I add a separate try catch around query compilation, that sets exception<X>Ptr  if
query compilation fails regardless of whether or not the --retry-with-try is set.

This way if the expression is wrapped in TRY we still verify that the exception is
consistent in the common and simplified paths.

(We could avoid re-executing the query with TRY if query compilation fails in the original
run, but I didn't implement that in the interest of keeping the code simple)

Reviewed By: pedroerp

Differential Revision: D34355097

fbshipit-source-id: bf616b251a6e85198c2dbf4525bc5cf8be4256bc