Add special interface for mapWriter when key and value are primitives. (#1036)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1036

When key and value are primitives, map can have std::like interface.
 `emplace` is enabled for that case.

Furthermore, to maximize performance, we want an interface that
skips the incremental increase of length and capacity check.

For that, `resize()` followed by subscript access can be used.
The map can be seen as std::Vector<std::tuple<k, v>>.
Examples of how the special interface can be used.

```

  mapWriter.resize(4);
  mapWriter[0] = std::make_tuple(1, 1);
  mapWriter[1] = std::make_tuple(2, std::nullopt);
  mapWriter[2] = std::make_tuple(3, std::optional<int64_t>{std::nullopt});
  mapWriter[3] = std::make_tuple(4, std::optional<int64_t>{11});

  mapWriter.emplace(1, 1);
  mapWriter.emplace(2, std::nullopt);
  mapWriter.emplace(3, std::optional<int64_t>{std::nullopt});
  mapWriter.emplace(4, std::optional<int64_t>{11});
```

Reviewed By: kevinwilfong

Differential Revision: D34217321

fbshipit-source-id: c39438dd238ef4753cc007e68271f4f5a1ac77ee