Skip execution of rows that have already thrown exceptions (#1081)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1081

Today, when we wrap an expression in a TRY we still evaluate expressions on rows that have already thrown an exception.

E.g. try(url_extract_path(to_base(4042309829,"c0"))) if to_base throws an exception, we'll still evaluate url_extract_path on that row

This can cause the system to crash with an SIGSEG for example because the argument passed in could have a null pointer as it was never initialized.

To address this I've updated evalAll to unselect any rows that have thrown exceptions (the TRY will turn the values in these rows to NULL later on anyway).

Reviewed By: pedroerp

Differential Revision: D34358383

fbshipit-source-id: 968ed13ef9eb602fb1c93d78b85535c61b9d0fdc