Add benchmark for try expressions (#1015)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1015

A customer was concerned that using TRY expressions might introduce a performance regression in cases where no exceptions were thrown (obviously throwing and catching exceptions has overhead).

I put together this benchmark to test the performance.

It sums 50 integer columns to create a deep tree of expressions, and runs it with no Try expression, a single Try expression around the entire sum, and a Try around each individual addition.

Also, just to measure the cost of exceptions, I compared dividing two columns, to dividing two columns where once the denominator was 0 (so one row threw an exception), to dividing a column by 0 (so every row threw an exception), all wrapped in TRY expressions.

## Results

```
============================================================================
velox/expression/benchmarks/TryBenchmark.cpp    relative  time/iter  iters/s
============================================================================
baseline                                                   277.71ns    3.60M
singleOuterTry                                             262.28ns    3.81M
nestedTries                                                280.72ns    3.56M
divideNoExceptions                                          16.55ns   60.43M
divideOneException                                          42.93ns   23.29M
divideAllExceptions                                         42.40us   23.59K
============================================================================
```

This shows no significant performance impact from the Try expressions, and the cost of a single exception is a 2x overhead (this is likely an upper bound given how simple the expression is), and >1000x worse when lots of exceptions are thrown (this seems to be dominated by the cost of C++'s exception handling, rather than any logic in the Try expression itself).

It's also worth noting that we log to stderr on exceptions which adds significantly to the cost, without that the overhead of a single exception is closer to 50%.

Reviewed By: laithsakka

Differential Revision: D34119788

fbshipit-source-id: 71eddeca2a798dc9e9c5e2ea3ae46044d99c01d4