Fixing TestServerInfoResource
Due to the DistributedQueryRunner creation within the test, it may end up taking longer than the timeout for the tests to finish. This leads to failures.
With the change, moving the DistributedQueryRunner creation outside of the method so the timeout will be actually accounted towards running the tests only
and not for the server startup.