Add join table reference check in JOIN ON clause
While it is permitted to omit the reference of the joining table in a
conditional join, the result of such join can lead to a cross join with
filter, which can harm the query execution performance. Often, such
queries are just a result of copy and paste from another join when
user updated the joining table name but forgot to update the join
condition.

This change will show the PERFORMANCE_WARNING to the user in cases
when JOIN ON clause misses the comparison expression with joining
table and other table.