Componetize SelectiveColumnReader (#1085)
Summary:
Split SelectiveColumnReader into format-specific readers and extract
generic logic for value processing,filtering and the column visitor
pattern.

The objective is reusing common pieces in selective readers for Parquet and Alpha.

The next step is adapting the Visitor signatures so that a single
Visitor can handle alternating direct and dictionary encodings in the
same stripe.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1085

Reviewed By: zzhao0

Differential Revision: D34400174

Pulled By: oerling

fbshipit-source-id: a16c2303a3b8704254d22a050090bb2065145d24