Add support for planning segmented aggregation
If the input of the aggregation is already pre-grouped by the prefix of
the GroupBy columns, we can flush the hash table whenever we see a new
pre-grouped key.

The execution side change is currently only available in the cpp stack.
This commit adds the planning side support to set the pre-grouped keys
in the AggregationNode plan node so that cpp stack can enable the
feature at execution time. The java stack will fallback to the
hash-based aggregation in this case.