Avoid Slice allocation for Statistics
StatisticsBuilder for String/Binary requires Slice input. Short lived
Slices are created from Blocks to compute statistics. This change
makes the statistics builder directly work with type, block and
position to avoid creation of short lived slices.

This change assumes that block.getSlice(position) is equivalent to
the type.getSlice(block, position). This assumption was already
made in the dictionary writer, this change expands it to all
Slice writers.