Write data directly to OutputBuffer
Before this change, a temporary slice is allocated for it to be
appended to the OrcOutputBuffer. For DirectColumnWriter this results
in large allocation of very short lived Slices. This change
adds a support for block to be directly written to the SliceOutput.

There are couple of functional changes.
1. If a type implements custom code for getSlice, this change will
break it. But ORC does not support other types. So this is not
possible.
2. This shifts some APIs from old APIs to similar APIs. If those APIs
have slightly different behavior, then this can cause regressions.