Removing scheduler hard coded request timeout value
When coordinator gets full GC pause which lasts more than 10 second, all scheduler requests to workers times out.
This result in retrying those requests. With 100s of tasks requests retry, causing memory pressure on coordinator if it's
running on low memory machine. Removing this config to make it configurable so we can tune it per cluster.