Add mechanism to provide context to exceptions (#1138)
Summary:
Errors happening during expression evaluation are hard to troubleshoot because
it is not clear which sub-expression triggered the error. This PR adds a
mechanism to specify context for exceptions via a thread_local variable and
uses it to specify the expression being evaluated and add it to the exception.

Here are some examples of the new "context" information being added to the
exceptions:

(c0 + c1) % 0 -> Context: mod(cast(plus(c0, c1)), literal)
c0 + (c1 % 0) -> Context: mod(cast(c1), literal)

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1138

Reviewed By: laithsakka, pedroerp

Differential Revision: D34704626

Pulled By: mbasmanova

fbshipit-source-id: b504b776ea43dca7f32d278611a8601ae5e0fd01