Clean up some unreachable code in the UnsafeRow24Deserializer
Summary:
Fixed-width types never go through `Deserialize` now, so renamed it to
`DeserializeVariableLength` and removed the fixed-width cases from the
type-switch. The next optimization would be to add optimized paths for
varchar/varbinary to row/array, so `DeserializeVariableLength` would only be
used for dispatching for map, struct, and row. At that point it might make
sense to simply inline it into DeserializeRow and DeserializeArray.

Reviewed By: miaoever

Differential Revision: D34807893

fbshipit-source-id: 46b1ef83cd3d4557b317797d8a45dc1a3da9fcea