Ensure we don't destroy Driver, which refers to a Task. (#1140)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1140

During creation of Drivers in the Task we can have a situation when the code throws and all created Drivers so far are being destroyed before their reference to Task is removed.
That causes crash due to DLOLG(FATAL) in debug mode. In release mode it would be fine.
We wrap driver creation into the try/catch to ensure that does not happen.

Reviewed By: mbasmanova

Differential Revision: D34706093

fbshipit-source-id: 67de5d33b1040d5af80820d1d2bdc4bef139fae4