Retry broadcast OOM with BHJ disabled within the same spark session
Presto on Spark uses temp storage for storing and distributing
broadcast tables. Spark driver performs the necessary threshold
checks on broadcast table and if the size is over the threshold,
the query fails with broadcast oom. The only way to fix this
failure is to disable broadcast join in the query.

As we are able to detect broadcast OOM on driver confidently,
we can just disable broadcast join, replan and resubmit the
query for execution. This can happen within the same spark
session itself and thus would not need any users intervention
for fixing such failures.