Remove Task::finishFuture method (#1203)
Summary:
Task::finishFuture() method is confusing as its name suggests that it returns a
future that is completed when task is finished. Instead, it returns a future
that completes when task is paused in response to Task::requestPause. This
change is to remove finishFuture() method and modify requestPause() method to
return a future that completes when task is paused.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1203

Reviewed By: kevinwilfong

Differential Revision: D34869908

Pulled By: mbasmanova

fbshipit-source-id: fc0d0b38e6b7048115c140b51cf2b91f50b3c717