Use and record writer time zone in ORC files
Cherry-pick of commit
https://github.com/trinodb/trino/pull/212/commits/f086b525d66f0a59f7b8fe60173f2e4eed000a28

Early versions of the Apache ORC writer made the mistake of recording
timestamps from an epoch that was relative to the time zone of the
writer. This was fixed in later versions by recording the writer time
zone in the stripe footer. Hive 3.1 always writes using UTC.

Presto used a global configuration for the writer time zone, which
was needed to handle old files, but was never updated to use the time
zone from the stripe footer.

On read, Presto now uses the stripe value if present, otherwise it
uses the configured value. On write, Presto continues to write
timestamps using the configured time zone, but now records this value
when writing files.

Co-authored-by: David Phillips <david@acz.org>