Modify DirectoryLister interface to pass partition information
In order to support mutation of partitions, we need partition information during listing time. We are changing the Directory lister interface for the same. We also need to inject additional split related information which has complicated structure and cannot be represented by the existing map structure in HiveSplit. We are going to use existing "extraFileInfo" field of type:byte[] and will leverage properties in customSplitInfo map for information necessary to deserialize the byte[].