Add BaseVector::estimateFlatSize method (#1225)
Summary:
BaseVector::estimateFlatSize method returns an estimate of the 'retainedSize' of
a flat representation of the data stored in this vector. This method is used to
report outputBytes statistics for individual operators.

Before this change outputBytes statistic was calculated using
BasedVector::retainedSize which caused double counting when operators
produced multiple dictionary vectors using the same base vector.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1225

Reviewed By: kgpai, oerling

Differential Revision: D34926559

Pulled By: mbasmanova

fbshipit-source-id: b8356beabda345845753102ab27a814acf6a0dd5