Add BaseVector::prepareForReuse method (#1231)
Summary:
To safely reuse a vector one needs to (1) ensure that the vector as well
as all its buffers and child vectors are singly-referenced and mutable
(for buffers); (2) clear append-only string buffers and child vectors
(elements of arrays, keys and values of maps, fields of structs).

The new static BaseVector::prepareForReuse method is designed to enable safe
vector reuse without too much boiler plate. It takes a 'vector' and a new size
and returns possibly new flat vector of the specified size that is safe to
reuse. If input 'vector' is not singly-referenced or not flat, returns a new
vector of the same type and specified size. If some of the buffers cannot be
reused, these buffers are reset. Child vectors are updated by calling this
method recursively with size zero.

Here is an example of usage:

```
void HashProbe::prepareOutput(vector_size_t size) {
  if (output_) {
    VectorPtr output = std::move(output_);
    BaseVector::prepareForReuse(output, size)
    output_ = std::static_pointer_cast<RowVector>(output);
  } else {
    output_ = std::static_pointer_cast<RowVector>(
        BaseVector::create(outputType_, size, pool()));
  }
}
```

Note that the vector is moved into VectorPtr variable before calling
BaseVector::prepareForReuse. This is necessary if the type of the vector is not
VectorPtr, e.g. RowVectorPtr. Passing a reference to RowVectorPtr will
increment reference count and effectively disable reuse as vector.unique
() check inside the BaseVector::prepareForReuse method will return false.

Update HashProbe and AssignUniqueId operators to use the new method.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1231

Reviewed By: oerling

Differential Revision: D34950824

Pulled By: mbasmanova

fbshipit-source-id: a3039c7b196a8b883e7591e1b1e38617c8c71df8