FileWriter to handle instrumentation (#1279)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1279

Add a templated wrapper class to handling writer instrumentation, which involves tracking the io stats from WS sink.

This wrapper can be applied outside of writer shim classes, as long as shim classes provide the specialization on how to extract some of the run stats through `OperationStatsExtractor`.

In addition to wiring up for trabant use cases. We also injected cpu measurement of flush time for velox dwrf writer, which is a more meaningful measurement than file close wall time. Flush time also allows us to more clearly calculate time spent on writes alone.

In order not to shrink current log retention, we print the serialized run stats instead of logging it to scuba. We'll populate and reuse some columns for a subset of the information.

Differential Revision: D35103858

fbshipit-source-id: 4ae97e71177aa8c10a38bb3573311cf0f1673706