Add Kll sketch implementation. (#1247)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1247

The current T-Digest implementation does not have a bounded
error, thus unable to provide a reliable guarantee for accuracy.  So
we want to reimplement the function using KLL sketch.  With this KLL sketch implementation, a proper k can guarantee the accuracy within 99% probability.

We did a comparison between T-Digest and KLL, and find for raw input, KLL has better performance for large input.  For merging intermediate results, T-Digest is faster.  Details can be found in `KllSketchBenchmark.cpp`.

Subsequent diffs will be created for serialization/deserialization and aggregation functions.

Reviewed By: mbasmanova

Differential Revision: D35009274

fbshipit-source-id: 034946fa00e78c751036b55accb9d1cb66875138