Fix Q1 and Q6 filters in TpchQueryBuilder (#1320)
Summary:
The queries in the TpchQueryBuilder build the TableScan node with an output
type containing the standard column names and use assignments to allow
flexibility in column names in the data files. Q1 and Q6 plans inside
TpchQueryBuilder are returning no results if the filter column name in the data
files differs from the standard column name. The issue is that the field names
specified in the filters provided to the TableScan node are not adjusted using
the assignments when building the ScanSpec. This PR fixes this by using the
data file names for the field names when building the filters.
ParquetTpchTest.cpp has been modified to test this change.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1320

Reviewed By: pedroerp

Differential Revision: D35283450

Pulled By: mbasmanova

fbshipit-source-id: 42badb55410dffb9b440a76e04a083cea428e0e2