Explicit handle of unknow type in FlatVector<StringView>::copy. (#1361)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1361

I want to add a check to asUnchecked function in BaseVector that is enabled in debug mode,
that make sure casting is only used when valid.

One of the tests was failing because FlatVector<StringView>::copy does an unsafe cast
(although it does not use the casted pointer when the input is null at a specific index).

This diff refactors the code to make explicit handling of source with type kind unknown.
For better future maintenance and code understanding. And also to enable the check added
in the previous diff.

Reviewed By: kevinwilfong

Differential Revision: D35408211

fbshipit-source-id: af86e0fe30d90da7c42078cfd57a03b5e84c8a43