Fix calculation error in deepInstanceSize of BlockBuilderStatus
Cherry-pick of https://github.com/trinodb/trino/pull/11710
Currently, in the deepInstanceSize function of BlockBuilderStatus
class, we recursively calculate the size of one BlockBuilderStatus
instance. However, we should exclude the static field which is not
the size for one instance, it's the size for the whole class and is
shared by each intance.

Co-authored-by: JackieTien97 <jackietien97@gmail.com>