Fix one case of fulfilling promises under a lock. (#1432)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1432

Fulfilling promises under a lock runs a risk of dealocking in case when the promise callback also uses some locks and is run in the same executor.
The common practice is to move out promises from class members into the local variables unders the lock and then fullfill promises, using local variables, outside of the lock.

Reviewed By: tanjialiang

Differential Revision: D35682204

fbshipit-source-id: 4e3c4519c5d07d75df1ae1b6028fa1f58ae11826