Fix bug when reapplying peeled dictionary encoding when nulls have been removed (#1026)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1026

When we call setDictionaryWrap to peel dictionary encoding, and then setWrapped to
reapply the dictionary encoding the size of the DictionaryVector can change.  This is
because we do not use the original vector size, but rather rows.end() for the size of the new
unpeeled vector.

This causes problems when nulls are removed, as when we call addNulls, the
DictionaryVector may be smaller than the number of rows when there were nulls at the end
of the batch.  In this case, addNulls attempts to resize the DictionaryVector which is not
supported.

To fix this, I changed addNulls in Expr.cpp to call setSize on the vector to effectively resize
it (all Vectors support setSize), and then call a new function ensureNullsCapacity which has
logic I moved from mutableNulls to resize nulls_ in the Vector (again all Vectors support
this).

I debated nulling out flatRawNulls_ in DictionaryVector when ensureNullsCapacity is called,
I decided not to since previously we could call mutableNulls on the DictionaryVector and
produce the same results without modifying it. In Exprs.cpp we call addNulls immediately
after which nulls out flatRawNulls_ anyway.

Reviewed By: mbasmanova

Differential Revision: D34176042

fbshipit-source-id: 896cba7fd07b2d184073ee63157ad303a348098b