Add dbgen wrapper to generate Vectors for `Nation` (#1422)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1422

Adding a wrapper around dbgen to generate TPC-H synthetic data and
return them as RowVector. For now only adding Nation tables, but as we agree on
the API I'll go ahead an add the other tables.
.
Note that for now I'm wrapping around duckDB's dbgen, but only using vanilla
dbgen. So in the next diffs we can add our own copy of dbgen (by vendoring it
or a submodule), and remove the duckDB dependency.

Reviewed By: mbasmanova

Differential Revision: D35634474

fbshipit-source-id: 892ec0efdbb93711866ba0b3e7ec5ba3c373465b