Replace coalesce function with a special form (#301)
Summary:
X-link: https://github.com/facebookresearch/torcharrow/pull/301

Coalesce needs to be a special form expression and cannot be a function because
it is not supposed to evaluate remaining inputs once there are no nulls. For
example, if column a has no nulls, the following expression should not try to
evaluate `b / 0` and should not fail.

```
coalesce(a, b / 0)
```

This change introduces CoalesceExpr special form and deletes coalesce function.

Also, fix a crash due to null pointer dereference in FlatVector::mutableRawValues.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1430

Reviewed By: laithsakka

Differential Revision: D35674128

Pulled By: mbasmanova

fbshipit-source-id: 2d16cb3404843322da67b699e7163d53757225bd