Remove transitionToRunning on Stage schedulingComplete()
The origin is `dead` code, since the check condition for each
task used SqlStageExecution#getState().
Stage could never lead to RUNNING by this condition.

We remove it, because it is safe and can reduce the task iteration
checks and code is clear(there is only one point to RUNNING)

Cherry-pick of https://github.com/trinodb/trino/pull/9004

Co-authored-by: guhanjie <guhanjiehao@163.com>