Making MappedMemory 'shared_from_this'. (#1497)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1497

X-link: https://github.com/facebookexternal/presto_cpp/pull/724

In order to extend the lifetime of MappedMemory when it is required, we inherit it from 'shared_from_this', so we can get a reference when we need.
This diff introduces the inheritance, makes sure we create MappedMemory and all its child classes using create_shared and finally, gets a reference to MappedMemory in StreamArena.

Reviewed By: tanjialiang

Differential Revision: D35989046

fbshipit-source-id: 7a60481d8ed73ec67e78eba0d1eb1c064d465322