Remove Stats and Jmx dependency from presto-orc
presto-orc for WriterStats used the OrcWriterFlushStats. This brings in
the JMX and stats dependency. Our team uses the orc as file reader and
writer. Due to recent introduction of drift dependency in airlift-stats
our binary size increased. Moving the OrcWriterStats to presto-hive
to avoid the dependency in presto-orc. In presto hive only class name
is used in the JMX export path, so this will have no change in the JMX.