Reuse memory more often when evaluating simple functions (#1521)
Summary:
When evaluating simple functions, check if one of the arguments can be re-used
for result. This is possible if
1. result type is fixed-width (e.g. primitive type except for VARCHAR and VARBINARY);
2. function doesn't produce nulls, e.g. has void call() or callNullable() and doesn't have callNullFree(),
3. argument and result types match,
4. argument is flat, uniquely-referenced and has uniquely-referenced values and nulls buffers.

This change reduces then number of memory allocations for a simple `(f - 0.5) * 2.0 + 0.3`
expression from 3 to 1. We evaluate 3 expressions:

* f1 = f - 0.5
* f2 = f1 * 2.0
* f3 = f2 + 0.3

We allocate new memory for f1, but then reuse it for f2 and f3. This can be visualized as:

* r = f - 0.5
* r = r * 2.0
* r = r + 0.3

Reusing argument for result is possible only if the processing loops over rows
and on each iteration reads the input, computes the result, then writes the
result. Here is a simple illustration:

```
std::vector<int> data;
for (int i = 0; i < data.size(); ++i) {
   data[i] = data[i]  + 5;
}
```

Requirements 1-4 above ensure that simple function evaluation matches this pattern.

For functions that produce nulls, the pattern of evaluation is different. It
involves (1) clearing nulls buffer of the result before the loop and(2) setting
null flags during the loop. This pattern is efficient for the case where null
results are very rare.

```
VectorPtr input;
VectorPtr result;

result.clearNulls();

for (int i = 0; i < input->size(); ++i) {
   auto result = func(input->valueAt(i));
   if (result is null) {
      result->setNull(i);
   } else {
      result->setValue(i);
   }
}
```

In this case, reusing input for result is not possible as `result.clearNulls()`
will modify the input before it has been processed. However, if input
doesn't have any nulls to begin with, it can be reused for result.

It is tricky to reuse input for results for variable-width functions, e.g.
VARCHAR. FlatVector<StringView> contains fixed-width values buffer and zero or
more variable-width string buffers that contain variable-width strings.
Updating an entry in FlatVector<StringView> vector modifies an entry in values
buffer and appends the new string to one of the string buffers that has space
or allocates a whole new string buffer. After the loop, the vector stores all
the old strings (input) and all the new strings in the string buffers. The
vector uses a lot more memory than necessary. To adjust for that effect one
needs to update the vector to remove the string buffers that contain strings
that are no longer referenced.

Similar challenge arise when considering reuse of arrays and maps where
modifying an entry involves appending to the elements, keys and values nested
vectors and requires cleanup after the loop.

Benchmark results:

```
gcc version 8.5.0 20210514 (Red Hat 8.5.0-10) (GCC)

model name      : Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz
cpu MHz         : 2999.998
cache size      : 25344 KB

====================================================================================================
velox_benchmark_basic_simple_arithmetic.json
====================================================================================================
    ✓ Pass: SimpleArithmetic.cpp/multiply                                  (1.12ns vs 1.11ns) +1.21%
    ✓ Pass: SimpleArithmetic.cpp/multiplySameColumn                        (0.57ns vs 0.56ns) +0.97%
    ✓ Pass: SimpleArithmetic.cpp/multiplyHalfNull                         (1.34ns vs 1.13ns) +15.15%
    ✓ Pass: SimpleArithmetic.cpp/multiplyConstant                         (0.67ns vs 0.59ns) +12.53%
    ✓ Pass: SimpleArithmetic.cpp/multiplyNested                           (7.64ns vs 1.79ns) +76.56%
    ✓ Pass: SimpleArithmetic.cpp/multiplyNestedDeep                     (18.87ns vs 13.02ns) +30.99%
    ✓ Pass: SimpleArithmetic.cpp/multiplyOutputVoid                       (1.17ns vs 1.00ns) +14.41%
    ✓ Pass: SimpleArithmetic.cpp/multiplyOutputNullable                   (1.15ns vs 1.02ns) +10.74%
    ✓ Pass: SimpleArithmetic.cpp/multiplyOutputAlwaysNull                  (1.52ns vs 1.49ns) +1.89%
    ✓ Pass: SimpleArithmetic.cpp/plusUnchecked                            (1.16ns vs 0.97ns) +16.53%
    ✓ Pass: SimpleArithmetic.cpp/plusChecked                              (3.56ns vs 3.19ns) +10.25%
  Pass: 44 (2 are faster):
    SimpleArithmetic.cpp/multiplyNested
    SimpleArithmetic.cpp/multiplyNestedDeep
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1521

Reviewed By: kgpai, kagamiori

Differential Revision: D36073380

Pulled By: mbasmanova

fbshipit-source-id: e7bf5a98040bd813949bfaab250b03d6f10c1e3f